+++ ft_nbrcount.c
@@ -3,18 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ft_nbrcount.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: elindber <elindber@student.42.fr>          +#+  +:+       +#+        */
+/*   By: elindber <elindber@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/03/11 14:23:51 by elindber          #+#    #+#             */
-/*   Updated: 2020/03/11 14:26:08 by elindber         ###   ########.fr       */
+/*   Updated: 2020/04/30 14:39:15 by elindber         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-uintmax_t	ft_nbrcount(intmax_t nbr)
+size_t	ft_nbrcount(intmax_t nbr)
 {
-	uintmax_t count;
+	size_t count;
 
 	count = 0;
 	if (nbr == 0)
